requirements.txt
easyocr==1.7.2
numpy==2.2.2
Pillow==10.1.0
Pillow==11.1.0
pynput==1.7.7
pyperclip==1.8.2
pyperclip==1.9.0
